Deciding how to deal with round data transfer
Originally, the idea was to have every tarkov round in meatov. H3 has less rounds than tarkov, meaning some tarkov rounds point to the same H3 round. This mean we can't get an H3 round's item data based on its H3ID alone because multiple tarkov ID might be using it. So we have to track rounds' item data in the various ammo containers. This adds an insane amount of complexity to the mod. My new idea is to make sure H3 rounds are only pointed by a single tarkov round. This will cause rounds to be missing from the game, but is acceptable considering the alternative.